# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manu local HTTP API",
    "version": "v1",
    "description": "Contrato HTTP local, sem autenticação, para o primeiro recorte de ingestão, consulta e inspeção de evidências. A instância é limitada ao endereço loopback e usa uma única organização configurada; o valor padrão é `local`. Os envelopes JSON têm versão `v1`. Erros usam `application/problem+json` e não expõem conteúdo bruto, segredos ou diagnósticos de provedores."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Servidor local padrão (o endereço configurado deve permanecer em loopback)."
    }
  ],
...
  "paths": {
...
    "/api/v1/ingestions": {
      "post": {
        "tags": ["ingestions"],
        "summary": "Aceita um bundle para ingestão assíncrona",
        "description": "O request deve ser `multipart/form-data` com as partes canônicas na ordem `manifest.json`, `artifacts.ndjson`, `contributions.ndjson` e, opcionalmente, `evidence.ndjson`. O job é aceito antes de seu processamento terminar. A organização é a única organização configurada na instalação e não é recebida no request.",
        "operationId": "createIngestion",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {"$ref": "#/components/schemas/AnalysisBundleMultipart"},
              "encoding": {
                "manifest.json": {"contentType": "application/json"},
                "artifacts.ndjson": {"contentType": "application/x-ndjson"},
                "contributions.ndjson": {"contentType": "application/x-ndjson"},
                "evidence.ndjson": {"contentType": "application/x-ndjson"}
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Bundle aceito; o job continua assíncrono.",
            "headers": {
              "X-Request-ID": {"$ref": "#/components/headers/RequestID"}
            },
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/IngestionResponse"},
                "example": {
                  "version": "v1",
                  "id": "11111111-1111-4111-8111-111111111111",
                  "organization_id": "local",
                  "state": "pending",
                  "stage": "validation",
                  "attempt_count": 0,
                  "counts": {"artifact_count": 0, "observation_count": 0, "evidence_count": 0, "failure_count": 0},
                  "request_id": "req_example_10"
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidIngestion"},
          "408": {"$ref": "#/components/responses/RequestTimeout"},
          "409": {"$ref": "#/components/responses/IngestionConflict"},
          "413": {"$ref": "#/components/responses/IngestionLimit"},
          "415": {"$ref": "#/components/responses/UnsupportedMedia"},
          "500": {"$ref": "#/components/responses/IngestionFailed"},
          "503": {"$ref": "#/components/responses/IngestionUnavailable"},
          "405": {"$ref": "#/components/responses/MethodNotAllowedPost"}
        }
      }
    },
    "/api/v1/ingestions/{ingestion_id}": {
      "parameters": [{"$ref": "#/components/parameters/IngestionID"}],
      "get": {
        "tags": ["ingestions"],
        "summary": "Consulta o estado de uma ingestão",
        "description": "O identificador deve ser um UUID textual de 36 caracteres. O retorno não inclui lease, erro bruto ou conteúdo secreto.",
        "operationId": "getIngestion",
        "responses": {
          "200": {
            "description": "Estado atual do job.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IngestionResponse"}}}
          },
          "400": {"$ref": "#/components/responses/InvalidIngestionPath"},
          "404": {"$ref": "#/components/responses/IngestionNotFound"},
          "408": {"$ref": "#/components/responses/RequestTimeout"},
          "500": {"$ref": "#/components/responses/IngestionFailed"},
          "503": {"$ref": "#/components/responses/IngestionUnavailable"},
          "405": {"$ref": "#/components/responses/MethodNotAllowedGet"}
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestID": {
        "description": "Identificador de correlação ecoado no corpo e na resposta. Se o request não o fornecer, o servidor gera um valor seguro.",
        "required": true,
        "schema": {
          "type": "string",
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9_.-]{1,128}$"
        },
        "example": "req_example_01"
      }
    },
    "parameters": {
      "IngestionID": {
        "name": "ingestion_id",
        "in": "path",
        "required": true,
        "description": "Identificador UUID textual da ingestão.",
        "schema": {"$ref": "#/components/schemas/UUID"},
        "example": "11111111-1111-4111-8111-111111111111"
      },
...
    },
    "responses": {
      "InvalidIngestion": {"$ref": "#/components/responses/Problem400Ingestion"},
      "InvalidIngestionPath": {"$ref": "#/components/responses/Problem400IngestionPath"},
      "IngestionNotFound": {"$ref": "#/components/responses/Problem404Ingestion"},
      "IngestionConflict": {"$ref": "#/components/responses/Problem409Ingestion"},
      "IngestionLimit": {"$ref": "#/components/responses/Problem413Ingestion"},
      "IngestionFailed": {"$ref": "#/components/responses/Problem500Ingestion"},
      "IngestionUnavailable": {"$ref": "#/components/responses/Problem503Ingestion"},
...
      "RequestTimeout": {
        "description": "O request foi cancelado ou excedeu o prazo; `code` é `request_canceled` ou `request_timeout`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "request_timeout", "message": "request timed out", "status": 408, "request_id": "req_example_30"}}}
      },
      "UnsupportedMedia": {
        "description": "O media type não é aceito pela operação.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "unsupported_media_type", "message": "content type is not supported", "status": 415, "request_id": "req_example_31"}}}
      },
      "MethodNotAllowedPost": {
        "description": "O método não é permitido para a coleção; `Allow: POST`.",
        "headers": {"Allow": {"schema": {"type": "string", "const": "POST"}}, "X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "method_not_allowed", "message": "method is not allowed", "status": 405, "request_id": "req_example_32"}}}
      },
      "MethodNotAllowedGet": {
        "description": "O método não é permitido para o recurso; `Allow: GET`.",
        "headers": {"Allow": {"schema": {"type": "string", "const": "GET"}}, "X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "method_not_allowed", "message": "method is not allowed", "status": 405, "request_id": "req_example_33"}}}
      },
      "Problem400Ingestion": {
        "description": "Bundle ou request de ingestão inválido (`invalid_ingestion`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "invalid_ingestion", "message": "ingestion payload is invalid", "status": 400, "request_id": "req_example_34"}}}
      },
      "Problem400IngestionPath": {
        "description": "O identificador ou caminho de ingestão é inválido (`invalid_ingestion_path`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "invalid_ingestion_path", "message": "ingestion path is invalid", "status": 400, "request_id": "req_example_35"}}}
      },
      "Problem404Ingestion": {
        "description": "Ingestão não encontrada (`ingestion_not_found`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "ingestion_not_found", "message": "ingestion was not found", "status": 404, "request_id": "req_example_36"}}}
      },
      "Problem409Ingestion": {
        "description": "Conflito de identidade da ingestão (`ingestion_conflict`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "ingestion_conflict", "message": "ingestion conflicts with existing work", "status": 409, "request_id": "req_example_37"}}}
      },
      "Problem413Ingestion": {
        "description": "Limite de body, bundle, manifest ou evidências excedido (`ingestion_limit_exceeded`); uma rejeição global anterior ao roteamento pode usar `body_too_large`. Padrões: body/bundle 64 MiB, manifest 1 MiB, stream de evidências 256 KiB, 10.000 unidades e 64 KiB por texto; os limites são configuráveis.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "ingestion_limit_exceeded", "message": "ingestion exceeds configured limits", "status": 413, "request_id": "req_example_38"}}}
      },
      "Problem500Ingestion": {
        "description": "Falha interna da ingestão (`ingestion_failed`).",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "ingestion_failed", "message": "ingestion failed", "status": 500, "request_id": "req_example_39"}}}
      },
      "Problem503Ingestion": {
        "description": "Serviço de ingestão indisponível ou ocupado (`ingestion_unavailable` ou `ingestion_busy`); o limite global de concorrência pode usar `concurrency_limited`.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}, "example": {"version": "v1", "code": "ingestion_unavailable", "message": "ingestion service is unavailable", "status": 503, "request_id": "req_example_40"}}}
      },
...
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
        "description": "UUID textual de 36 caracteres; a validação do handler aceita maiúsculas e minúsculas."
      },
...
      "RequestID": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9_.-]{1,128}$"
      },
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "code", "message", "status", "request_id"],
        "properties": {
          "version": {"type": "string", "const": "v1"},
          "code": {"type": "string", "pattern": "^[a-z0-9_.-]{1,64}$"},
          "message": {"type": "string", "maxLength": 128},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "request_id": {"$ref": "#/components/schemas/RequestID"}
        },
        "description": "Envelope de erro seguro. O texto é fixo e não carrega conteúdo do request, DSN, segredo ou mensagem de provedor."
      },
      "IngestionCounts": {
        "type": "object",
        "additionalProperties": false,
        "required": ["artifact_count", "observation_count", "evidence_count", "failure_count"],
        "properties": {
          "artifact_count": {"type": "integer", "minimum": 0},
          "observation_count": {"type": "integer", "minimum": 0},
          "evidence_count": {"type": "integer", "minimum": 0},
          "failure_count": {"type": "integer", "minimum": 0}
        }
      },
      "IngestionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "id", "organization_id", "state", "stage", "attempt_count", "counts", "request_id"],
        "properties": {
          "version": {"type": "string", "const": "v1"},
          "id": {"$ref": "#/components/schemas/UUID"},
          "organization_id": {"type": "string", "minLength": 1},
          "state": {"type": "string", "enum": ["pending", "running", "completed", "partial", "failed"]},
          "stage": {"type": "string", "enum": ["validation", "canonical_persistence", "textual_projection", "relational_projection", "embedding_projection", "activation"]},
          "attempt_count": {"type": "integer", "minimum": 0},
          "counts": {"$ref": "#/components/schemas/IngestionCounts"},
          "diagnostic_code": {"type": "string", "pattern": "^[a-z0-9_.-]{1,64}$"},
          "request_id": {"$ref": "#/components/schemas/RequestID"}
        }
      },
      "AnalysisBundleMultipart": {
        "type": "object",
        "additionalProperties": false,
        "required": ["manifest.json", "artifacts.ndjson", "contributions.ndjson"],
        "properties": {
          "manifest.json": {"type": "string", "format": "binary", "contentMediaType": "application/json", "description": "Parte obrigatória; máximo padrão de 1 MiB."},
          "artifacts.ndjson": {"type": "string", "format": "binary", "contentMediaType": "application/x-ndjson", "description": "Parte obrigatória de artefatos."},
          "contributions.ndjson": {"type": "string", "format": "binary", "contentMediaType": "application/x-ndjson", "description": "Parte obrigatória de contribuições."},
          "evidence.ndjson": {"type": "string", "format": "binary", "contentMediaType": "application/x-ndjson", "description": "Parte opcional quando o bundle fornece evidências; stream máximo padrão de 256 KiB e 10.000 unidades."}
        },
        "description": "O servidor também limita o body/bundle a 64 MiB por padrão. A ordem e os nomes são parte do contrato; os limites podem ser configurados."
      },
...
    }
  }
}